+++ libs/render_delegate/instancer.cpp
@@ -96,11 +96,13 @@ void HdArnoldInstancer::Sync(HdSceneDelegate* sceneDelegate, HdRenderParam* rend
 // Sample a primvar, check that the keys have the correct number of instances otherwise get only the sample at the keyframe
 // We have to do this because hydra SamplePrimvar
 template <typename VectorT>
+
 static void SamplePrimvarChecked(
     HdSceneDelegate* delegate, const SdfPath& id, const TfToken& key, const GfVec2f& shutterRange, VectorT& out)
 {
     HdArnoldSampledPrimvarType sample;
     delegate->SamplePrimvar(id, key, &sample);
+
     if (sample.count >= 1) {
         // We expect SamplePrimvar to return the same number of elements in sampled arrays.
         // However this number might be different than the number of element at the frame.
@@ -114,6 +116,7 @@ static void SamplePrimvarChecked(
             }
         }
     }
+
     HdArnoldUnboxResample(sample, shutterRange, out);
 }
 
@@ -144,17 +147,13 @@ void HdArnoldInstancer::_SyncPrimvars(HdDirtyBits dirtyBits, HdArnoldRenderParam
                 continue;
             }
             if (primvar.name == GetInstanceTransformsToken()) {
-                SamplePrimvarChecked(
-                    GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _transforms);
+                SamplePrimvarChecked(GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _transforms);
             } else if (primvar.name == GetRotateToken()) {
-                SamplePrimvarChecked(
-                    GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _rotates);
+                SamplePrimvarChecked(GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _rotates);
             } else if (primvar.name == GetScaleToken()) {
-                SamplePrimvarChecked(
-                    GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _scales);
+                SamplePrimvarChecked(GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _scales);
             } else if (primvar.name == GetTranslateToken()) {
-                SamplePrimvarChecked(
-                    GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _translates);
+                SamplePrimvarChecked(GetDelegate(), id, primvar.name, renderParam->GetShutterRange(), _translates);
             } else {
                 HdArnoldInsertPrimvar(
                     _primvars, primvar.name, primvar.role, primvar.interpolation, GetDelegate()->Get(id, primvar.name),
@@ -163,28 +162,26 @@ void HdArnoldInstancer::_SyncPrimvars(HdDirtyBits dirtyBits, HdArnoldRenderParam
         }
     }
 
-    // Note Cyril: it shouldn't be necessary to mark the instancer clean as it is done by hydra
+    // NOTE: it shouldn't be necessary to mark the instancer clean as it is done later on by hydra
     changeTracker.MarkInstancerClean(id);
 }
 
 void HdArnoldInstancer::ResampleInstancePrimvars()
 {
     const auto& id = GetId();
-    // Compute the number of instances expected as we need to check the size of the array primvars later on
+    std::lock_guard<std::mutex> lock(_mutex);
+    // Recompute the sampled primvars only if they were previously sampled
     if (_transforms.count) {
-        SamplePrimvarChecked(
-            GetDelegate(), id, GetInstanceTransformsToken(), _samplingInterval, _transforms);
+        SamplePrimvarChecked(GetDelegate(), id, GetInstanceTransformsToken(), _samplingInterval, _transforms);
     }
     if (_rotates.count) {
-        SamplePrimvarChecked(
-            GetDelegate(), id, GetRotateToken(), _samplingInterval, _rotates);
+        SamplePrimvarChecked(GetDelegate(), id, GetRotateToken(), _samplingInterval, _rotates);
     }
     if (_scales.count) {
         SamplePrimvarChecked(GetDelegate(), id, GetScaleToken(), _samplingInterval, _scales);
     }
     if (_translates.count) {
-        SamplePrimvarChecked(
-            GetDelegate(), id, GetTranslateToken(), _samplingInterval, _translates);
+        SamplePrimvarChecked(GetDelegate(), id, GetTranslateToken(), _samplingInterval, _translates);
     }
 }
 
@@ -193,8 +190,8 @@ void HdArnoldInstancer::CalculateInstanceMatrices(HdArnoldRenderDelegate* render
 {
     const SdfPath& instancerId = GetId();
     HdArnoldRenderParam * renderParam = reinterpret_cast<HdArnoldRenderParam*>(renderDelegate->GetRenderParam());
-    
-    // If the sampling interval has changed
+
+    // If the sampling interval has changed we need to resample the translate, orientations and scales
     if (UpdateSamplingInterval(renderParam->GetShutterRange())){
         ResampleInstancePrimvars();
     }
@@ -214,10 +211,15 @@ void HdArnoldInstancer::CalculateInstanceMatrices(HdArnoldRenderDelegate* render
     // most samples and use its time range.
     // TODO(pal): Improve this further by using the widest time range and calculate sample count based on that.
     _AccumulateSampleTimes(instancerTransforms, sampleArray);
-    _AccumulateSampleTimes(_transforms, sampleArray);
-    _AccumulateSampleTimes(_translates, sampleArray);
-    _AccumulateSampleTimes(_rotates, sampleArray);
-    _AccumulateSampleTimes(_scales, sampleArray);
+    {
+        // Another mesh can be resampling the instances primvars, we need to lock 
+        std::lock_guard<std::mutex> lock(_mutex);
+        _AccumulateSampleTimes(_transforms, sampleArray);
+        _AccumulateSampleTimes(_translates, sampleArray);
+        _AccumulateSampleTimes(_rotates, sampleArray);
+        _AccumulateSampleTimes(_scales, sampleArray);
+    }
+
 
     // By default _deformKeys will take over sample counts
     if (sampleArray.count <= 2 && _deformKeys < 2 && _deformKeys > -1 ) {