+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 
 ## Next Feature release
 
+### Features
+
+- [usd#2264](https://github.com/Autodesk/arnold-usd/issues/2264) - Skip translation of invisible primitives in the render delegate
+- [usd#2277](https://github.com/Autodesk/arnold-usd/issues/2277) - Ensure child nodes are properly destroyed in the hydra procedural
+- [usd#2276](https://github.com/Autodesk/arnold-usd/issues/2276) - Improve default interactive FPS settings in the render delegate
+
+## Next Bugfix release
+
+### Bug fixes
+
+- [usd#2269](https://github.com/Autodesk/arnold-usd/issues/2269) - Connections are not processed correctly during hydra procedural updates
+- [usd#2268](https://github.com/Autodesk/arnold-usd/issues/2268) - Reenable instancer motion blur on the first rendered frame by recomputing the transform matrices when the shutter is updated.
+
+## [7.4.1.0] - 2025-03-26
+
 ### Bug fixes
 - [usd#2254](https://github.com/Autodesk/arnold-usd/issues/2254) - Fix bug happening when the number of normal keys is less than the number of point keys.
 - [usd#2231](https://github.com/Autodesk/arnold-usd/issues/2231) - Fix motion blur when an instancer as a varying number of instances.